requirements.txt
# requirements.txt
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.2,<5.0.0
langchain>=0.1.16,<0.2.0
openai>=1.30.2,<2.0.0
chromadb>=0.4.24,<0.5.0
tiktoken>=0.7.0,<1.0.0
pydantic>=2.7.1,<3.0.0
numpy>=1.26.4,<2.0.0
python-dotenv>=1.0.0,<2.0.0
flask>=3.0.2,<4.0.0
langchain-openai>=0.1.2,<0.2.0